requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-cors==4.0.0
torch>=2.0.0
transformers>=4.30.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0 